backend/requirements.txt
alembic==1.4.0
aniso8601==7.0.0
argon2-cffi==19.2.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
dnspython==1.16.0
eventlet==0.25.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-GraphQL==2.0.1
Flask-JWT-Extended==3.24.1
Flask-Migrate==2.5.2
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
gevent-websocket==0.10.1
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-iot==0.3.1
googleapis-common-protos==1.51.0
graphene==2.1.8
graphene-sqlalchemy==2.2.2
graphql-core==2.3.1
graphql-relay==2.0.1
graphql-server-core==1.2.0
greenlet==0.4.15
grpc-google-iam-v1==0.12.3
grpcio==1.27.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.18.1
pandas==1.0.1
promise==2.3
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
python-engineio==3.11.2
python-http-client==3.2.5
python-socketio==4.4.0
pytz==2019.3
requests==2.23.0
rsa==4.0
Rx==1.6.1
sendgrid==6.1.2
singledispatch==3.4.0.3
six==1.14.0
SQLAlchemy==1.3.13
SQLAlchemy-Utils==0.36.1
urllib3==1.25.8
Werkzeug==1.0.0
